// typescript/user-interface/packages/weavess-core/src/ts/types/types.ts
export interface TimeRange {
  /** Start Time in seconds */
  startTimeSecs: number;

  /** End Time in seconds */
  endTimeSecs: number;
}